feat: Added functions in post ata and fixed bugs

+++ sid/src/main/java/br/sc/weg/sid/model/entities/PropostasLog.java
@@ -20,7 +20,7 @@ public class PropostasLog {
     private String demandaTituloPropostaLog;
 
     @Column(name = "demandaTempoExecucaoPropostaLog", nullable = false)
-    private Integer demandaTempoExecucaoPropostaLog;
+    private Long demandaTempoExecucaoPropostaLog;
 
     @Column(name = "demandaValorPropostaLog", nullable = false)
     private Double demandaValorPropostaLog;
